refactor: Update ViewTransition imports path
