+++ packages/hsp/ui/src/modules/guesart/layout/index.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react'
-import ViewTransition from '@hsp/ui/src/components/app/ViewTransition'
+import ViewTransition from '@hsp/ui/src/utils/react/view-transition'
 import { ButtonLink } from '@hsp/ui/src/components/base/button'
 import { ThemeToggle } from '@hsp/ui/src/components/theme/ThemeToggle'
 import NoSsr from '@hsp/ui/src/components/app/NoSsr'
