+++ packages/hsp/ui/src/modules/guesart/pages/lobby/index.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react'
 
-import ViewTransition from '@hsp/ui/src/components/app/ViewTransition'
+import ViewTransition from '@hsp/ui/src/utils/react/view-transition'
 import { Card } from '@hsp/ui/src/components/base/card'
 import { FormGroup } from './group'
