+++ packages/hsp/ui/src/test/vitest.setup.ts
@@ -2,7 +2,7 @@ import { vi } from 'vitest'
 import '@testing-library/jest-dom/vitest'
 import { PropsWithChildren } from 'react'
 
-vi.mock('@hsp/ui/src/components/app/ViewTransition', () => {
+vi.mock('@hsp/ui/src/utils/react/view-transition', () => {
   return {
     default: ({ children }: PropsWithChildren) => children,
   }
