+++ packages/hsp/ui/src/utils/react/view-transition.tsx
